Replace react-hot-loader with react-refresh
`react-hot-loader` is now deprecated. The recommendation of the authors is to use `react-refresh`.

I've installed that package and the Webpack plugin and set it up easily. I also refactor the example code a little bit to show case some of the stuff.

Some caveats, the root `App` component has no hot loading now. Bummer. But React Refresh is much more stable and will be the defacto solution now.

I've also set up the `style-loader` thingy in dev, due to the fact that it works better with `react-refresh`. This inline the styles in the `head` of the document so it's easier to reload everything when it happens.