Upgrade webpack to v5 (#150)
Upgrade all dependencies and loaders to Webpack 5.

This brings improved build performance, better and smaller bundles.

We only needed to tweak the webpack config a little bit to make the upgrade (just the devtool option)

Update: for some reason the build started to fail on lint. We had to include all the extra eslint plugins because `yarn` won't hoist the tools installed from our config. This fixes it also.