Replace webpack with snowpack (#231)
Snowpack is a much lighter alternative to Webpack, which features an entirely new codebase and revitalized internals.

The motivations are pretty simple, we want to move away from transpiling/bundling hell into the native ES Modules support. Snowpack is leading the way on that.

This PR ensures we have a clean, Typescript ready codebase. It also features a very trimmed down PostCSS config, we just `postcss-nested` for nested rules and queries like SASS and `custom-media-properties` to define reusable media queries easily.

We still use Webpack, but it's completely abstracted by Snowpack. For production, we need to bundle the stuff, so I resorted to their recommended way of doing that.

The objective of this PR is ensuring people start using some "new" things in the JS space. Snowpack can and will crash with some old npm packages, and it's a feature (to me at least) so we can start using modern stuff instead of old bloat. Without radical change, no progress can be made.