Add a frontend to demonstrate the API auth (#280)
Simple frontend that sets up route based async imports, private routes (that redirect unauthenticated user).

Installed `react-router-dom` to demonstrate that.

Also added Wallaby tests to assert the functionality.

Also added to change the cors config. We need to specify the origin to include credentials from another domain. On the dev server, our API server is at localhost:4000 and the frontend is at localhost:8080. On production, they are served from the same domain. In cases where we want to host frontend and backend, we need to either use JWT tokens on storage or specify more allowed origins.