Use foreman to start the development server (#313)
Our `bin/server` script caused some formatting issues and was a bit brittle in regards to zombie node processed and the like. For example, if the `elixir` process crashed, the script would exit and you would have a zombie process (snowpack) even with the traps and the like. And it wasn't that cross platform.

Like our `rails-starter` we decided to use foreman. Both linux and mac users can easily install it.

This way you get a nice way of running both the snowpack dev server and iex without issues.