Fix crash when visiting sign up after error (#314)
After the **login** page gets an error, if you visit the sign-up page you crash the app. This happens because the sign-up page views the `error` state and tries to get validation errors from that, but the error is a login error, which is an empty response with 401 code.

This PR adds a fix so that whenever the route of the app changes, we reset the error state if it exists, to avoid extra renders.

This PR also fixes an issue with code splitting. The way we had it all the chunks were being evaluated at first. We just need to change the `importPath` prop to a function so it gets evaluated correctly by `loadable`.