Set the frontend API_URL through a env variable (#335)
The new snowpack version has a bug regarding NODE_ENV visibility on the build (it's always development). Therefore our mechanism of defining the base API url was not working correctly.

I suggested a fix on Snowpack, which is merged and is awaiting release: https://github.com/snowpackjs/snowpack/pull/3132

Anyway, I think that by setting the API_URL via an env variable, we can have more flexibility, deploying the frontend assets to a CDN even, and pointing it to the location of the backend.